feat: rewrite with TypeScript and hooks
To provide a better experience for RNGlobalize users that are also using TypeScript, RNGlobalize has
been rewritten in TypeScript, providing up-to-date typings builded with the package. Additionally,
the library has moved to using hooks, which simplifies many of the component implementations.

BREAKING CHANGE: The `withGlobalize` HOC has been removed in favor of using the `useContext` hook
and the context is now exported as `GlobalizeContext` instead of `Context`.

closes #47